Replace tabs with spaces

+++ development-setup-wizard.mjs
@@ -53,9 +53,9 @@ async function checkPrereqs() {
 
 
 async function createEnvFiles() {
-	if (process.env.CAMPHORIC_SETUP_WIZARD_RECREATE_ENV === 'no') {
-		return;
-	}
+  if (process.env.CAMPHORIC_SETUP_WIZARD_RECREATE_ENV === 'no') {
+    return;
+  }
 
   const dir = './env';
   let envDir;
@@ -67,15 +67,15 @@ async function createEnvFiles() {
   }
 
   if (envDir.isDirectory()) {
-		if (process.env.CAMPHORIC_SETUP_WIZARD_RECREATE_ENV === undefined) {
-			const { blowaway } = await inquirer.prompt([{
-				name: 'blowaway',
-				type: 'confirm',
-				message: `it looks like '${dir}' already exists, do you want to blow it away and start over?`,
-			}]);
+    if (process.env.CAMPHORIC_SETUP_WIZARD_RECREATE_ENV === undefined) {
+      const { blowaway } = await inquirer.prompt([{
+        name: 'blowaway',
+        type: 'confirm',
+        message: `it looks like '${dir}' already exists, do you want to blow it away and start over?`,
+      }]);
 
-			if (!blowaway) return;
-		}
+      if (!blowaway) return;
+    }
 
     await fs.rm(dir, { recursive: true, force: true });
   }
@@ -152,9 +152,9 @@ function dockerComposeUp() {
 
 // for now, no customization, we're keeping it simple
 async function setEnvValues() {
-	if (process.env.CAMPHORIC_SETUP_WIZARD_RECREATE_SECRET_KEY === 'no') {
-		return;
-	}
+  if (process.env.CAMPHORIC_SETUP_WIZARD_RECREATE_SECRET_KEY === 'no') {
+    return;
+  }
 
   const filename = './env/local/django.env';
   let contents = await fs.readFile(filename, 'utf-8').then(c => c.toString().trim());
@@ -172,7 +172,7 @@ async function setEnvValues() {
     contents += '\nSECRET_KEY=your-django-secret-key-here';
   }
   
-	if (currentKey !== 'your-django-secret-key-here' && process.env.CAMPHORIC_SETUP_WIZARD_RECREATE_SECRET_KE === undefined) {
+  if (currentKey !== 'your-django-secret-key-here' && process.env.CAMPHORIC_SETUP_WIZARD_RECREATE_SECRET_KE === undefined) {
     const { replaceKey } = await inquirer.prompt([{
       name: 'replaceKey',
       type: 'confirm',
@@ -187,9 +187,9 @@ async function setEnvValues() {
   contents = contents.replace(/SECRET_KEY=.*/, `SECRET_KEY='${secretKey}'`);
   await fs.writeFile(filename, contents, 'utf-8');
   hooray(`Successfully wrote secret key to ${filename}`);
-	
-	// Rebuild if we added secret key
-	await dockerComposeUp();
+  
+  // Rebuild if we added secret key
+  await dockerComposeUp();
 }
 
 async function getSecretKey() {
